Avoid bundling graphql/execution/execute into apollo-link-schema.
Importing graphql/execution/execute adds about 30KB to the minified
footprint of a typical application (6.8KB after gzip). This estimate was
obtained using the https://github.com/apollographql/react-apollo/pull/2839
example application, which is bundled with Rollup.

This is a significant pill to swallow under any circumstances, but before
this commit apollo-link-schema was bundling all of that code into itself,
preventing other packages from sharing the execution code, which very
likely led to its duplication in some applications.

With this commit, the self-size of apollo-link-schema falls from 19.3 KB
to just 398 bytes (minified+gzip), which means we can finally stop
excluding it from the filesize check.
